% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spotify.R
\name{convert_spaces}
\alias{convert_spaces}
\title{Converting spaces in a string to '%20}
\usage{
convert_spaces(string)
}
\arguments{
\item{string}{A string}
}
\value{
A string.
}
\description{
Converting spaces in a string to '%20
}
