% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_auth.R
\name{spotify_auth}
\alias{spotify_auth}
\title{Authenticate with Spotify's API}
\usage{
spotify_auth()
}
\value{
`NULL`
}
\description{
A function to allow the user to authenticate with Spotify's API. This will refresh expired tokens.
In order to interact with the Spotify API, please be sure to create an app within their system. See additional references to get started.
}
\seealso{
Get started with the Spotify API here: https://developer.spotify.com/documentation/web-api/quick-start/
}
