% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_search.R
\name{spotify_search}
\alias{spotify_search}
\title{Search with Spotify's API}
\usage{
spotify_search(track_name = NULL, artist_name = NULL, album_name = NULL)
}
\arguments{
\item{name}{A string. The name of what the user wants to search for within Spotify's database}

\item{type}{A string. The scope of what the search is limited to. Can be one or multiple of "album", "track", "artist"}
}
\value{

}
\description{
Search with Spotify's API
}
