% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_search.R
\name{spotify_search}
\alias{spotify_search}
\title{Title}
\usage{
spotify_search(name, type)
}
\arguments{
\item{name}{A String. The name of what I want to search for within Spotify's database.}

\item{type}{A String/Vector. The scope of what my search is limited to. Can be one or multiple of "album", "track", "artist".}
}
\value{
A Data Frame.
}
\description{
Title
}
